requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
black==25.12.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distlib==0.4.0
distro==1.9.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.1
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
identify==2.6.16
idna==3.11
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
librt==0.7.8
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.0
openai==2.14.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pathspec==1.0.3
pdfminer.six==20251230
pdfplumber==0.11.9
pgvector==0.4.2
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
pypdfium2==5.3.0
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
ruff==0.14.13
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
selenium==4.39.0
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
sqlparse==0.5.4
sympy==1.14.0
tenacity==9.1.2
-e git+https://github.com/zxSUPHASANxz/THE_ONE_FINAL.git@f29e7b62d204a7e031c1e7efdc380673a863f687#egg=the_one_v3
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
virtualenv==20.36.1
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.2
